test: Move getLineOrRoute tests
